+++ storage/src/adb/benches/any_init.rs
@@ -7,7 +7,7 @@ use commonware_runtime::{
     Runner as _, ThreadPool,
 };
 use commonware_storage::{
-    adb::any::{Any, Config as AConfig},
+    adb::any::fixed::{Any, Config as AConfig},
     translator::EightCap,
 };
 use criterion::{criterion_group, Criterion};
